% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{compute_reactive_in_list}
\alias{compute_reactive_in_list}
\title{Compute reactive value in list}
\usage{
compute_reactive_in_list(options)
}
\arguments{
\item{options}{The list to calculate values over}
}
\value{
The list of the calculated values
}
\description{
Given a list of items, calculate the actual value of the item in the list. It is a helper function to get the value of shiny::reactive value, and then need to be use in a reactive function. If the item is an R litteral, then this function returns it as it is.
}
