% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_to_dataframe.R
\name{EnvironmentList_to_dataframe}
\alias{EnvironmentList_to_dataframe}
\title{Transform an homgenous environment list to data.frame}
\usage{
EnvironmentList_to_dataframe(environmentList, ...)
}
\arguments{
\item{environmentList}{An environment list}

\item{...}{A list of name}
}
\value{
The dataset build upon the given environment's name subset. If omitted, this subset is build by all names referencing a primitive datatype
}
\description{
Extract the given name from all the environment in the list, and return it as a data.frame. If no names are given, extract all the primitive bounded variable
}
