% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_to_dataframe.R
\name{extract_from_environment}
\alias{extract_from_environment}
\title{Transform an environment to a vector}
\usage{
extract_from_environment(environment, ...)
}
\arguments{
\item{environment}{An environment}

\item{...}{A list of name}
}
\value{
The vector build upon the given environment's name subset.
}
\description{
Extract the given names field from the environment as a vector.
}
