% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_to_dataframe.R
\name{get_attributes}
\alias{get_attributes}
\title{Return the primitive attributes of an environment}
\usage{
get_attributes(environment)
}
\arguments{
\item{environment}{An environment}
}
\value{
The name's subset of the environment primitive variable's type
}
\description{
Return the primitive attributes of an environment
}
