% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_annotation_explorer.R
\name{mod_annotation_explorer_ui}
\alias{mod_annotation_explorer_ui}
\title{annotation_explorer UI Function}
\usage{
mod_annotation_explorer_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{id}{Environment id of the module}
}
\description{
A shiny::selectInput wrapper around AnnotationsApi.search_annotations (\code{\link[opensilexClientToolsR:AnnotationsApi]{ search_annotations }})
}
