% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_authentification.R
\name{mod_authentification_ui}
\alias{mod_authentification_ui}
\title{authentification UI Function}
\usage{
mod_authentification_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
The UI part of the module handling the connection to an opensilex instance. Required for all other module of this package.
}
