% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_device_explorer.R
\name{mod_device_explorer_ui}
\alias{mod_device_explorer_ui}
\title{device_explorer UI Function}
\usage{
mod_device_explorer_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{id}{Environment id of the module}
}
\description{
A shiny::selectInput wrapper around DevicesApi.search_devices (\code{\link[opensilexClientToolsR:DevicesApi]{ search_devices }})
}
