% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_document_explorer.R
\name{mod_document_explorer_ui}
\alias{mod_document_explorer_ui}
\title{document_explorer UI Function}
\usage{
mod_document_explorer_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{id}{Environment id of the module}
}
\description{
A shiny::selectInput wrapper around DocumentsApi.search_documents (\code{\link[opensilexClientToolsR:DocumentsApi]{ search_documents }})
}
