% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_event_explorer.R
\name{mod_event_explorer_server}
\alias{mod_event_explorer_server}
\title{event_explorer Server Functions}
\usage{
mod_event_explorer_server(
  id,
  authentification_module,
  api_function_options = list(),
  widget_options = list()
)
}
\arguments{
\item{id}{Internal parameter for {shiny}.}

\item{authentification_module}{Authentification module from this package (\code{\link{mod_authentification_server}})}

\item{api_function_options}{List of options to pass to EventsApi.search_events from opensilexClientToolsR package (\code{\link[opensilexClientToolsR:EventsApi]{ search_events }})}

\item{widget_options}{List of options to pass to \code{\link[shiny]{selectInput}}}
}
\value{
A named list with various reactive values
\describe{
 \item{input}{The module input, to be used for binding events}
 \item{options}{The options used to call EventsApi.search_events }
 \item{selected}{The selection of the shiny::selectInput widget }
 \item{choices}{A named list containing the selection choices} 
 \item{result_df}{A dataframe containing the results of EventsApi.search_events call}
}
}
\description{
event_explorer Server Functions
}
