% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_event_explorer.R
\name{mod_event_explorer_ui}
\alias{mod_event_explorer_ui}
\title{event_explorer UI Function}
\usage{
mod_event_explorer_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{id}{Environment id of the module}
}
\description{
A shiny::selectInput wrapper around EventsApi.search_events (\code{\link[opensilexClientToolsR:EventsApi]{ search_events }})
}
