% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_experiment_explorer.R
\name{mod_experiment_explorer_ui}
\alias{mod_experiment_explorer_ui}
\title{experiment_explorer UI Function}
\usage{
mod_experiment_explorer_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{id}{Environment id of the module}
}
\description{
A shiny::selectInput wrapper around ExperimentsApi.search_experiments (\code{\link[opensilexClientToolsR:ExperimentsApi]{ search_experiments }})
}
