% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_facilitie_explorer.R
\name{mod_facilitie_explorer_ui}
\alias{mod_facilitie_explorer_ui}
\title{facilitie_explorer UI Function}
\usage{
mod_facilitie_explorer_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{id}{Environment id of the module}
}
\description{
A shiny::selectInput wrapper around OrganisationsApi.search_infrastructure_facilities (\code{\link[opensilexClientToolsR:OrganisationsApi]{ search_infrastructure_facilities }})
}
