% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_factor_explorer.R
\name{mod_factor_explorer_ui}
\alias{mod_factor_explorer_ui}
\title{factor_explorer UI Function}
\usage{
mod_factor_explorer_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{id}{Environment id of the module}
}
\description{
A shiny::selectInput wrapper around FactorsApi.search_factors (\code{\link[opensilexClientToolsR:FactorsApi]{ search_factors }})
}
