% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_germplasm_explorer.R
\name{mod_germplasm_explorer_ui}
\alias{mod_germplasm_explorer_ui}
\title{germplasm_explorer UI Function}
\usage{
mod_germplasm_explorer_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{id}{Environment id of the module}
}
\description{
A shiny::selectInput wrapper around GermplasmApi.search_germplasm (\code{\link[opensilexClientToolsR:GermplasmApi]{ search_germplasm }})
}
