% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_group_explorer.R
\name{mod_group_explorer_ui}
\alias{mod_group_explorer_ui}
\title{group_explorer UI Function}
\usage{
mod_group_explorer_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{id}{Environment id of the module}
}
\description{
A shiny::selectInput wrapper around SecurityApi.search_groups (\code{\link[opensilexClientToolsR:SecurityApi]{ search_groups }})
}
