% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_position_explorer.R
\name{mod_position_explorer_server}
\alias{mod_position_explorer_server}
\title{position_explorer Server Functions}
\usage{
mod_position_explorer_server(
  id,
  authentification_module,
  api_function_options = list(),
  widget_options = list()
)
}
\arguments{
\item{id}{Internal parameter for {shiny}.}

\item{authentification_module}{Authentification module from this package (\code{\link{mod_authentification_server}})}

\item{api_function_options}{List of options to pass to PositionsApi.search_position_history from opensilexClientToolsR package (\code{\link[opensilexClientToolsR:PositionsApi]{ search_position_history }})}

\item{widget_options}{List of options to pass to \code{\link[shiny]{selectInput}}}
}
\value{
A named list with various reactive values
\describe{
 \item{input}{The module input, to be used for binding events}
 \item{options}{The options used to call PositionsApi.search_position_history }
 \item{selected}{The selection of the shiny::selectInput widget }
 \item{choices}{A named list containing the selection choices} 
 \item{result_df}{A dataframe containing the results of PositionsApi.search_position_history call}
}
}
\description{
position_explorer Server Functions
}
