% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_position_explorer.R
\name{mod_position_explorer_ui}
\alias{mod_position_explorer_ui}
\title{position_explorer UI Function}
\usage{
mod_position_explorer_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{id}{Environment id of the module}
}
\description{
A shiny::selectInput wrapper around PositionsApi.search_position_history (\code{\link[opensilexClientToolsR:PositionsApi]{ search_position_history }})
}
