% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scientific_object_explorer.R
\name{mod_scientific_object_explorer_ui}
\alias{mod_scientific_object_explorer_ui}
\title{scientific_object_explorer UI Function}
\usage{
mod_scientific_object_explorer_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{id}{Environment id of the module}
}
\description{
A shiny::selectInput wrapper around ScientificObjectsApi.search_scientific_objects (\code{\link[opensilexClientToolsR:ScientificObjectsApi]{ search_scientific_objects }})
}
