% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specie_explorer.R
\name{mod_specie_explorer_ui}
\alias{mod_specie_explorer_ui}
\title{specie_explorer UI Function}
\usage{
mod_specie_explorer_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{id}{Environment id of the module}
}
\description{
A shiny::selectInput wrapper around SpeciesApi.get_all_species (\code{\link[opensilexClientToolsR:SpeciesApi]{ get_all_species }})
}
