% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_subclasse_explorer.R
\name{mod_subclasse_explorer_ui}
\alias{mod_subclasse_explorer_ui}
\title{subclasse_explorer UI Function}
\usage{
mod_subclasse_explorer_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{id}{Environment id of the module}
}
\description{
A shiny::selectInput wrapper around OntologyApi.search_sub_classes_of (\code{\link[opensilexClientToolsR:OntologyApi]{ search_sub_classes_of }})
}
