% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tree_explorer.R
\name{mod_tree_explorer_ui}
\alias{mod_tree_explorer_ui}
\title{tree_explorer UI Function}
\usage{
mod_tree_explorer_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{id}{Environment id of the module}
}
\description{
A shiny::selectInput wrapper around OrganisationsApi.search_infrastructures_tree (\code{\link[opensilexClientToolsR:OrganisationsApi]{ search_infrastructures_tree }})
}
