% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_variable_detail_explorer.R
\name{mod_variable_detail_explorer_ui}
\alias{mod_variable_detail_explorer_ui}
\title{variable_detail_explorer UI Function}
\usage{
mod_variable_detail_explorer_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{id}{Environment id of the module}
}
\description{
A shiny::selectInput wrapper around VariablesApi.search_variables_details (\code{\link[opensilexClientToolsR:VariablesApi]{ search_variables_details }})
}
